requirements.txt
alembic==0.7.6
attrs==15.0.0
blinker==1.3
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.4.0
Flask-Principal==0.4.0
Flask-Script==2.0.5
Flask-Security==1.7.4
Flask-SQLAlchemy==2.0
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
passlib==1.6.2
SQLAlchemy==1.0.4
Werkzeug==0.10.4
WTForms==2.0.2
